jbFundAccessConstraints vars to uint256 and add checks

+++ contracts/structs/JBFundAccessConstraints.sol
@@ -7,11 +7,11 @@ struct JBFundAccessConstraints {
   // The terminal within which the distribution limit and the overflow allowance applies.
   IJBTerminal terminal;
   // The amount of the distribution limit.
-  uint248 distributionLimit;
+  uint256 distributionLimit;
   // The currency that the distribution limit are denoted in.
-  uint8 distributionLimitCurrency;
+  uint256 distributionLimitCurrency;
   // The amount of the allowance.
-  uint248 overflowAllowance;
+  uint256 overflowAllowance;
   // The currency that the overflow allowance are denoted in.
-  uint8 overflowAllowanceCurrency;
+  uint256 overflowAllowanceCurrency;
 }
